ADD(ui/view): Job Details And Search Results
Add Search View And Job Details View

Implemented a job details page. Some bugs with rendering
the qualifications and requirements still.

Implemented a search page. Enabling users to filter which
kind of jobs they are looking for. This page works fine
for getting results. But needs to extend the cards to trigger
going to that job's details.